pyproject.toml
[project]
name = "multi-agent-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "aiosqlite==0.21.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.13.0",
    "appdirs==1.4.4",
    "attrs==26.1.0",
    "backoff==2.2.1",
    "banks==2.4.1",
    "bcrypt==5.0.0",
    "build==1.4.2",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.6",
    "chromadb==1.1.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "crewai==1.12.2",
    "cryptography==46.0.6",
    "dataclasses-json==0.6.7",
    "deprecated==1.3.1",
    "deprecation==2.1.0",
    "dirtyjson==1.0.8",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "docstring-parser==0.17.0",
    "durationpy==0.10",
    "et-xmlfile==2.0.0",
    "filelock==3.25.2",
    "filetype==1.2.0",
    "flatbuffers==25.12.19",
    "frozenlist==1.8.0",
    "fsspec==2026.3.0",
    "googleapis-common-protos==1.73.1",
    "greenlet==3.3.2",
    "griffe==2.0.2",
    "griffecli==2.0.2",
    "griffelib==2.0.2",
    "grpcio==1.78.0",
    "h11==0.16.0",
    "h2==4.3.0",
    "hf-xet==1.4.2",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==1.8.0",
    "hyperframe==6.1.0",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "importlib-resources==6.5.2",
    "instructor==1.14.5",
    "jinja2==3.1.6",
    "jiter==0.11.1",
    "joblib==1.5.3",
    "json-repair==0.25.3",
    "json5==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.1.1",
    "jsonref==1.1.0",
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "kubernetes==35.0.0",
    "lance-namespace==0.6.1",
    "lance-namespace-urllib3-client==0.6.1",
    "lancedb==0.30.0",
    "langchain==1.2.13",
    "langchain-core==1.2.23",
    "langchain-openai==1.1.12",
    "langchain-text-splitters>=1.1.1",
    "langgraph==1.1.3",
    "langgraph-checkpoint==4.0.1",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.12",
    "langsmith==0.7.22",
    "linkify-it-py==2.1.0",
    "llama-index==0.14.19",
    "llama-index-cli==0.5.6",
    "llama-index-core==0.14.19",
    "llama-index-embeddings-openai==0.6.0",
    "llama-index-instrumentation==0.5.0",
    "llama-index-llms-openai==0.7.4",
    "llama-index-vector-stores-qdrant==0.10.0",
    "llama-index-workflows==2.17.1",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "marshmallow==3.26.2",
    "mcp==1.26.0",
    "mdit-py-plugins==0.5.0",
    "mdurl==0.1.2",
    "mmh3==5.2.1",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.6.1",
    "nltk==3.9.4",
    "numpy==2.4.3",
    "oauthlib==3.3.1",
    "onnxruntime==1.24.4",
    "openai==2.30.0",
    "openpyxl==3.1.5",
    "opentelemetry-api==1.34.1",
    "opentelemetry-exporter-otlp-proto-common==1.34.1",
    "opentelemetry-exporter-otlp-proto-grpc==1.34.1",
    "opentelemetry-exporter-otlp-proto-http==1.34.1",
    "opentelemetry-proto==1.34.1",
    "opentelemetry-sdk==1.34.1",
    "opentelemetry-semantic-conventions==0.55b1",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "overrides==7.7.0",
    "packaging==26.0",
    "pdfminer-six==20251230",
    "pdfplumber==0.11.9",
    "pillow==12.1.1",
    "platformdirs==4.9.4",
    "portalocker==2.7.0",
    "posthog==5.4.0",
    "propcache==0.4.1",
    "protobuf==5.29.6",
    "pyarrow==23.0.1",
    "pybase64==1.4.3",
    "pycparser==3.0",
    "pydantic==2.11.10",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pyjwt==2.12.1",
    "pypdfium2==5.6.0",
    "pypika==0.51.1",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.22",
    "pywin32==311",
    "pyyaml==6.0.3",
    "qdrant-client==1.17.1",
    "referencing==0.37.0",
    "regex==2026.1.15",
    "requests==2.33.0",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.3.3",
    "rpds-py==0.30.0",
    "setuptools==82.0.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.48",
    "sse-starlette==3.3.3",
    "starlette==1.0.0",
    "sympy==1.14.0",
    "tenacity==9.1.4",
    "textual==8.2.0",
    "tiktoken==0.12.0",
    "tinytag==2.2.1",
    "tokenizers==0.22.2",
    "tomli==2.0.2",
    "tomli-w==1.1.0",
    "tqdm==4.67.3",
    "typer==0.23.1",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "uc-micro-py==2.0.0",
    "urllib3==2.6.3",
    "uuid-utils==0.14.1",
    "uv==0.9.30",
    "uvicorn==0.42.0",
    "watchfiles==1.1.1",
    "websocket-client==1.9.0",
    "websockets==16.0",
    "wrapt==2.1.2",
    "xxhash==3.6.0",
    "yarl==1.23.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
